requirements.txt
streamlit>=1.30
tensorflow==2.20.0
numpy
opencv-python-headless
pillow
